refactor: update ContainerChildItem to use entry prop and simplify drop handling in MainArea
