+++ src/components/MainArea.vue
@@ -14,11 +14,7 @@
     </div>
     <div class="entry-panel" ref="entryPanelRef">
       <ContainerChildItem
-        :children="children"
-        :drop-allowed="dropAllowed"
-        @drop="onDrop"
-        @dragover="onDragOver"
-        @remove="removeChild"
+        :entry="mainContainer"
       />
     </div>
     <div class="connection-panel">
@@ -27,8 +23,7 @@
 </template>
 
 <script>
-import { ref, computed } from 'vue'
-import { useDroppable } from '../composables/useDroppable'
+import { ref } from 'vue'
 import { useEntryOperation } from '../composables/useEntryOperation'
 import { entryState } from '../composables/useEntryState'
 import { useEntryRect } from '../composables/useEntryRect'
@@ -41,65 +36,10 @@ export default {
   },
   
   setup() {
-    // Get composable
-    const {
-      isDroppable,
-      onDragOver,
-      onDrop,
-      setOnDropCallBack
-    } = useDroppable()
-    const { 
-      addBlock,
-      addContainer,
-      removeEntry,
-      reorderEntry,
-      moveEntry
-    } = useEntryOperation()
+    const { addContainer } = useEntryOperation()
 
     // Create a top-level container & register it in EntryManager
     const mainContainer = addContainer(null, 'main-area', 0)
-    
-    /**
-     * Remove a child entry
-     * @param {string} id - ID of the child to remove
-     */
-    const removeChild = (id) => {
-      removeEntry(id)
-    }
-    
-    // Set custom callbacks for drop event
-    setOnDropCallBack((event, index) => {
-      // Get data directly from event.dataTransfer
-      const entryType = event.dataTransfer.getData('entryType')
-      const entryName = event.dataTransfer.getData('entryName')
-      const entryId = event.dataTransfer.getData('entryId')
-      const sourceId = event.dataTransfer.getData('sourceId')
-      
-      if (!entryId) {
-        // Create and insert a new element
-        if (index !== null) {
-          if (entryType === 'block') {
-            addBlock(mainContainer.id, entryName, index)
-          } else if (entryType === 'container') {
-            addContainer(mainContainer.id, entryName, index)
-          }
-        }
-      } else {
-        if (!sourceId || sourceId === mainContainer.id) {
-          // Reorder within MainArea
-          reorderEntry(mainContainer.id, entryId, index)
-        } else {
-          // Drag & drop from a container
-          moveEntry(entryId, mainContainer.id, index)
-        }
-      }
-    })
-
-    // Array of children
-    const children = computed(() => mainContainer.children)
-
-    // For determining whether to allow the drop
-    const dropAllowed = isDroppable(mainContainer.id)
 
     // Template ref for the entry panel
     const entryPanelRef = ref(null)
@@ -107,11 +47,7 @@ export default {
 
     // Return values and methods to use in <template>
     return {
-      onDragOver,
-      onDrop,
-      removeChild,
-      children,
-      dropAllowed,
+      mainContainer,
       entryState,
       entryPanelRef,
       entryLayoutMap
